move BTF info query to ebpf module
