+++ pkg/ebpf/co_re_unsupported.go
@@ -3,13 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2024-present Datadog, Inc.
 
-//go:build linux_bpf
+//go:build !linux_bpf
 
-package modules
+package ebpf
 
-import "github.com/DataDog/datadog-agent/pkg/ebpf"
+import "errors"
 
-// GetBTFLoaderInfo Returns where the ebpf BTF files were sourced from, only on linux
+// GetBTFLoaderInfo is not supported without linux_bpf
 func GetBTFLoaderInfo() (string, error) {
-	return ebpf.GetBTFLoaderInfo()
+	return "", errors.New("GetBTFLoaderInfo not supported without linux_bpf")
 }
