+++ pkg/process/net/common_linux.go
@@ -32,7 +32,7 @@ const (
 	telemetryURL         = "http://unix/telemetry"
 	conntrackCachedURL   = "http://unix/" + string(sysconfig.NetworkTracerModule) + "/debug/conntrack/cached"
 	conntrackHostURL     = "http://unix/" + string(sysconfig.NetworkTracerModule) + "/debug/conntrack/host"
-	ebpfBTFLoaderURL     = "http://unix/" + string(sysconfig.NetworkTracerModule) + "/debug/ebpf_btf_loader"
+	ebpfBTFLoaderURL     = "http://unix/" + string(sysconfig.EBPFModule) + "/btf_loader_info"
 	netType              = "unix"
 )
